backend/requirements.txt
# FastAPI y servidor
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Base de datos
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
pymongo>=4.6.0
motor>=3.3.2

# Autenticación y seguridad
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
PyJWT>=2.8.0

# Validación de datos y emails
pydantic>=2.5.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# NLP y Machine Learning
transformers>=4.35.2
torch>=2.0.0
sentencepiece>=0.1.99
protobuf>=4.25.1

# Cliente HTTP
requests>=2.31.0
httpx>=0.25.2

# Scheduler para tareas automáticas
APScheduler==3.10.4

# SendGrid para emails
sendgrid>=6.11.0

# Testing
pytest>=7.4.3